+++ packages/genesys/packages/react-components/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.0.25
 
-[Package.json] Update dependecies: react-components-core & react-utils
+[Package.json] Update dependecies: react-components-core
 [Config] Add locale to the config
 
 ## 3.0.2